+++ benchmark/benches/bench_query_posvel_soa.jl
@@ -1,14 +1,17 @@
 
 function setup_query_posvel_soa(n_entities::Int)
-    world = World(PositionSoA, VelocitySoA)
+    world = World(
+        Position => StructArrayStorage,
+        Velocity => StructArrayStorage,
+    )
     for i in 1:n_entities
-        new_entity!(world, (PositionSoA(i, i * 2), VelocitySoA(1, 1)))
+        new_entity!(world, (Position(i, i * 2), Velocity(1, 1)))
     end
-    for (_, pos_column, vel_column) in @Query(world, (PositionSoA, VelocitySoA))
+    for (_, pos_column, vel_column) in @Query(world, (Position, Velocity))
         for i in eachindex(pos_column)
             @inbounds pos = pos_column[i]
             @inbounds vel = vel_column[i]
-            @inbounds pos_column[i] = PositionSoA(pos.x + vel.dx, pos.y + vel.dy)
+            @inbounds pos_column[i] = Position(pos.x + vel.dx, pos.y + vel.dy)
         end
     end
     return world
@@ -16,11 +19,11 @@ end
 
 function benchmark_query_posvel_soa(args, n)
     world = args
-    for (_, pos_column, vel_column) in @Query(world, (PositionSoA, VelocitySoA))
+    for (_, pos_column, vel_column) in @Query(world, (Position, Velocity))
         for i in eachindex(pos_column)
             @inbounds pos = pos_column[i]
             @inbounds vel = vel_column[i]
-            @inbounds pos_column[i] = PositionSoA(pos.x + vel.dx, pos.y + vel.dy)
+            @inbounds pos_column[i] = Position(pos.x + vel.dx, pos.y + vel.dy)
         end
     end
     return world