+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Features
 
 - Adds an event system with events for structural changes as well as custom events (#155)
-- Adds storage modes Vector and StructArray, which can be selected per component type (#207)
+- Adds storage modes Vector and StructArray, which can be selected per component type (#207, #225)
 - Adds field view, so field access for query columns works equally for Vector and StructArray storages (#213)
 
 ### Other